refactor : ParticleHelper.cpp으로 옮김.
